Move test class setup to separate python file
