+++ tests/test_phosim.py
@@ -24,29 +24,15 @@ import sys
 import unittest
 
 import lsst.utils.tests
-from lsst.utils import getPackageDir
 from lsst.afw.geom import arcseconds, Extent2I
 import lsst.afw.image
 import lsst.obs.base.tests
 
+from lsst.obs.lsst.testHelper import ObsLsstButlerTests
 
-class TestPhosim(lsst.obs.base.tests.ObsTests, lsst.utils.tests.TestCase):
 
-    @classmethod
-    def tearDownClass(cls):
-        cls._mapper._LsstCamMapper__clearCache()
-        del cls._mapper
-        del cls._butler
-
-    @classmethod
-    def setUpClass(cls):
-        product_dir = getPackageDir('obs_lsst')
-        cls.data_dir = os.path.join(product_dir, 'data', 'input', 'phosim')
-
-        cls._butler = lsst.daf.persistence.Butler(root=cls.data_dir)
-        mapper_class = cls._butler.getMapperClass(root=cls.data_dir)
-        mapper_class._LsstCamMapper__clearCache()
-        cls._mapper = mapper_class(root=cls.data_dir)
+class TestPhosim(lsst.obs.base.tests.ObsTests, ObsLsstButlerTests):
+    instrumentDir = "phosim"
 
     def setUp(self):
         dataIds = {'raw': {'visit': 204595, 'detectorName': 'S20', 'raftName': 'R11'},
